Move last remaining methods to util.ts
