+++ packages/client/src/rpc/modules/engine/util.ts
@@ -1,10 +1,23 @@
 import { Block } from '@ethereumjs/block'
-import { bigIntToHex, bytesToHex, equalsBytes } from '@ethereumjs/util'
+import { Hardfork } from '@ethereumjs/common'
+import {
+  BIGINT_1,
+  bigIntToHex,
+  bytesToHex,
+  bytesToUnprefixedHex,
+  equalsBytes,
+  hexToBytes,
+} from '@ethereumjs/util'
+
+import { UNSUPPORTED_FORK } from '../../error-code'
+
+import { type BlobsBundleV1, type ChainCache, type PayloadStatusV1, Status } from './types'
 
 import type { Chain } from '../../../blockchain'
 import type { BlobsBundle } from '../../../miner'
-import type { BlobsBundleV1, ChainCache } from './types'
+import type { ExecutionPayloadBodyV1 } from './types'
 import type { ExecutionPayload } from '@ethereumjs/block'
+import type { Common } from '@ethereumjs/common'
 
 /**
  * Formats a block to {@link ExecutionPayloadV1}.
@@ -146,5 +159,141 @@ export const validExecutedChainBlock = async (
     return null
   } catch (error: any) {
     return null
+  }
+}
+
+/**
+ * Returns the block hash as a 0x-prefixed hex string if found valid in the blockchain, otherwise returns null.
+ */
+export const validHash = async (
+  hash: Uint8Array,
+  chain: Chain,
+  chainCache: ChainCache
+): Promise<string | null> => {
+  const { remoteBlocks, executedBlocks, invalidBlocks, skeleton } = chainCache
+  const maxDepth = chain.config.engineParentLookupMaxDepth
+
+  try {
+    let validParent: Block | null = null
+    for (let inspectedParents = 0; inspectedParents < maxDepth; inspectedParents++) {
+      const unPrefixedHashStr = bytesToUnprefixedHex(hash)
+      validParent =
+        remoteBlocks.get(unPrefixedHashStr) ??
+        (await skeleton.getBlockByHash(hash, true)) ??
+        (await chain.getBlock(hash))
+
+      // if block is invalid throw error and respond with null validHash
+      if (invalidBlocks.get(unPrefixedHashStr) !== undefined) {
+        throw Error(`References an invalid ancestor`)
+      }
+
+      // if block is executed the return with this hash
+      const isBlockExecuted =
+        (executedBlocks.get(unPrefixedHashStr) ?? (await validExecutedChainBlock(hash, chain))) !==
+        null
+      if (isBlockExecuted) {
+        return bytesToHex(hash)
+      } else {
+        hash = validParent.header.parentHash
+      }
+    }
+  } catch (_error: any) {
+    // ignore error thrown by the loop and return null below
+  }
+
+  // if we are here, either we can't find valid parent till maxDepth or the ancestor was invalid
+  // or there was a lookup error. in all these instances return null
+  return null
+}
+
+/**
+ * Validates that the block satisfies post-merge conditions.
+ */
+export const validateTerminalBlock = async (block: Block, chain: Chain): Promise<boolean> => {
+  const ttd = chain.config.chainCommon.hardforkTTD(Hardfork.Paris)
+  if (ttd === null) return false
+  const blockTd = await chain.getTd(block.hash(), block.header.number)
+
+  // Block is terminal if its td >= ttd and its parent td < ttd.
+  // In case the Genesis block has td >= ttd it is the terminal block
+  if (block.isGenesis()) return blockTd >= ttd
+
+  const parentBlockTd = await chain.getTd(block.header.parentHash, block.header.number - BIGINT_1)
+  return blockTd >= ttd && parentBlockTd < ttd
+}
+
+/**
+ * Returns a block from a payload.
+ * If errors, returns {@link PayloadStatusV1}
+ */
+export const assembleBlock = async (
+  payload: ExecutionPayload,
+  chain: Chain,
+  chainCache: ChainCache
+): Promise<{ block?: Block; error?: PayloadStatusV1 }> => {
+  const { blockNumber, timestamp } = payload
+  const { config } = chain
+  const common = config.chainCommon.copy()
+
+  // This is a post merge block, so set its common accordingly
+  // Can't use setHardfork flag, as the transactions will need to be deserialized
+  // first before the header can be constucted with their roots
+  const ttd = common.hardforkTTD(Hardfork.Paris)
+  common.setHardforkBy({ blockNumber, td: ttd !== null ? ttd : undefined, timestamp })
+
+  try {
+    const block = await Block.fromExecutionPayload(payload, { common })
+    // TODO: validateData is also called in applyBlock while runBlock, may be it can be optimized
+    // by removing/skipping block data validation from there
+    await block.validateData()
+    return { block }
+  } catch (error) {
+    const validationError = `Error assembling block from payload: ${error}`
+    config.logger.error(validationError)
+    const latestValidHash = await validHash(hexToBytes(payload.parentHash), chain, chainCache)
+    const response = {
+      status: `${error}`.includes('Invalid blockHash') ? Status.INVALID_BLOCK_HASH : Status.INVALID,
+      latestValidHash,
+      validationError,
+    }
+    return { error: response }
+  }
+}
+
+export const getPayloadBody = (block: Block): ExecutionPayloadBodyV1 => {
+  const transactions = block.transactions.map((tx) => bytesToHex(tx.serialize()))
+  const withdrawals = block.withdrawals?.map((wt) => wt.toJSON()) ?? null
+
+  return {
+    transactions,
+    withdrawals,
+  }
+}
+
+export function validateHardforkRange(
+  chainCommon: Common,
+  methodVersion: number,
+  checkNotBeforeHf: Hardfork | null,
+  checkNotAfterHf: Hardfork | null,
+  timestamp: bigint
+) {
+  if (checkNotBeforeHf !== null) {
+    const hfTimeStamp = chainCommon.hardforkTimestamp(checkNotBeforeHf)
+    if (hfTimeStamp !== null && timestamp < hfTimeStamp) {
+      throw {
+        code: UNSUPPORTED_FORK,
+        message: `V${methodVersion} cannot be called pre-${checkNotBeforeHf}`,
+      }
+    }
+  }
+
+  if (checkNotAfterHf !== null) {
+    const nextHFTimestamp = chainCommon.nextHardforkBlockOrTimestamp(checkNotAfterHf)
+    if (nextHFTimestamp !== null && timestamp >= nextHFTimestamp) {
+      throw {
+        code: UNSUPPORTED_FORK,
+        message: `V${methodVersion + 1} MUST be called post-${checkNotAfterHf}`,
+      }
+    }
   }
 }
